styling fixes and add top tab component for navigation to /usage/per-project
