+++ jsapp/js/components/common/tabs.module.scss
@@ -3,7 +3,7 @@
 
 .root {
   width: auto;
-  padding-top: sizes.$x40;
+  padding-top: sizes.$x20;
   margin: 0 sizes.$x40;
   display: flex;
   flex-direction: row;
@@ -22,7 +22,7 @@
 
   &.active {
     color: colors.$kobo-dark-blue;
-    background-color: colors.$kobo-light-blue;
+    background-color: colors.$kobo-bg-blue;
     border-bottom: sizes.$x2 solid colors.$kobo-dark-blue;
     border-radius: sizes.$x8 sizes.$x8 0 0;
   }
